# lab4/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
        "/token": {
            "post": {
                "summary": "Login For Access Token",
                "operationId": "login_for_access_token_token_post",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_login_for_access_token_token_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{login}/send_message/{target_login}": {
            "put": {
                "summary": "Send Message",
                "operationId": "send_message_users__login__send_message__target_login__put",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "login",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Login"
                        }
                    },
                    {
                        "name": "target_login",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Target Login"
                        }
                    },
                    {
                        "name": "body",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Body"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Chat"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{login}/chat/{target_login}": {
            "get": {
                "summary": "Get Chat",
                "operationId": "get_chat_users__login__chat__target_login__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "login",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Login"
                        }
                    },
                    {
                        "name": "target_login",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Target Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Chat"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{login}/wall": {
            "get": {
                "summary": "Get Wall",
                "operationId": "get_wall_users__login__wall_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "login",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Wall"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{login}/create_post": {
            "post": {
                "summary": "Create Post",
                "operationId": "create_post_users__login__create_post_post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "login",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Login"
                        }
                    },
                    {
                        "name": "post_txt",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Post Txt"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Wall"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{name}/{surname}": {
            "get": {
                "summary": "Get User By Name",
                "operationId": "get_user_by_name_users__name___surname__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Name"
                        }
                    },
                    {
                        "name": "surname",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Surname"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User-Output"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{login}": {
            "get": {
                "summary": "Get User",
                "operationId": "get_user_users__login__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "login",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User-Output"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "summary": "Create User",
                "operationId": "create_user_users_post",
                "parameters": [
                    {
                        "name": "current_user",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Current User"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User-Input"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User-Output"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Body_login_for_access_token_token_post": {
                "properties": {
                    "grant_type": {
                        "anyOf": [
                            {
                                "type": "string",
                                "pattern": "password"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Grant Type"
                    },
                    "username": {
                        "type": "string",
                        "title": "Username"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    },
                    "scope": {
                        "type": "string",
                        "title": "Scope",
                        "default": ""
                    },
                    "client_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Client Id"
                    },
                    "client_secret": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Client Secret"
                    }
                },
                "type": "object",
                "required": [
                    "username",
                    "password"
                ],
                "title": "Body_login_for_access_token_token_post"
            },
            "Chat": {
                "properties": {
                    "messages": {
                        "items": {
                            "$ref": "#/components/schemas/Message"
                        },
                        "type": "array",
                        "title": "Messages"
                    },
                    "login1": {
                        "type": "string",
                        "title": "Login1"
                    },
                    "login2": {
                        "type": "string",
                        "title": "Login2"
                    }
                },
                "type": "object",
                "required": [
                    "messages",
                    "login1",
                    "login2"
                ],
                "title": "Chat"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "Message": {
                "properties": {
                    "body": {
                        "type": "string",
                        "title": "Body"
                    },
                    "user_name": {
                        "type": "string",
                        "title": "User Name"
                    }
                },
                "type": "object",
                "required": [
                    "body",
                    "user_name"
                ],
                "title": "Message"
            },
            "User-Input": {
                "properties": {
                    "login": {
                        "type": "string",
                        "title": "Login"
                    },
                    "hashed_password": {
                        "type": "string",
                        "title": "Hashed Password"
                    },
                    "name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Name"
                    },
                    "surname": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Surname"
                    },
                    "wall": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/Wall"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "chats": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/Chat"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Chats"
                    }
                },
                "type": "object",
                "required": [
                    "login",
                    "hashed_password"
                ],
                "title": "User"
            },
            "User-Output": {
                "properties": {
                    "login": {
                        "type": "string",
                        "title": "Login"
                    },
                    "hashed_password": {
                        "type": "string",
                        "title": "Hashed Password"
                    },
                    "name": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Name"
                    },
                    "surname": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Surname"
                    },
                    "wall": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/Wall"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "chats": {
                        "anyOf": [
                            {
                                "items": {
                                    "$ref": "#/components/schemas/Chat"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Chats"
                    }
                },
                "type": "object",
                "required": [
                    "login",
                    "hashed_password"
                ],
                "title": "User"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
            "Wall": {
                "properties": {
                    "posts": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Posts"
                    }
                },
                "type": "object",
                "required": [
                    "posts"
                ],
                "title": "Wall"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "token"
                    }
                }
            }
        }
    }
}